+++ internal/api/sdsubmit_test.go
@@ -105,26 +105,28 @@ func Test_SDSubmit_GetFiles_Fail(t *testing.T) {
 
 func Test_SDSubmit_GetFiles_Pass(t *testing.T) {
 	// Mock
-	testFile := []file{
+	testFile := []fileInfo{
 		{
 			FileID:                "file1",
 			DatasetID:             "dataset1",
 			DisplayFileName:       "file1.txt",
+			FilePath:              "dataset1/file1.txt",
 			FileName:              "file1.txt",
 			FileSize:              10,
 			DecryptedFileSize:     10,
 			DecryptedFileChecksum: "abc123",
-			FileStatus:            "READY",
+			Status:                "READY",
 		},
 		{
 			FileID:                "file2",
 			DatasetID:             "dataset2",
 			DisplayFileName:       "file2.txt",
+			FilePath:              "dataset2/file2.txt",
 			FileName:              "file2.txt",
 			FileSize:              10,
 			DecryptedFileSize:     10,
 			DecryptedFileChecksum: "abc123",
-			FileStatus:            "PENDING",
+			Status:                "PENDING",
 		},
 	}
 	testFileJSON, _ := json.Marshal(testFile)
@@ -157,16 +159,17 @@ func Test_SDSubmit_GetFiles_Pass(t *testing.T) {
 
 func Test_SDSubmit_GetFiles_Split_Pass(t *testing.T) {
 	// Mock
-	testFile := []file{
+	testFile := []fileInfo{
 		{
 			FileID:                "file1",
 			DatasetID:             "dataset1",
 			DisplayFileName:       "file1.txt",
+			FilePath:              "dataset1/file1.txt",
 			FileName:              "file1.txt",
 			FileSize:              20,
 			DecryptedFileSize:     20,
 			DecryptedFileChecksum: "if6ox",
-			FileStatus:            "ready",
+			Status:                "ready",
 		},
 	}
 	testFileJSON, _ := json.Marshal(testFile)
@@ -414,14 +417,6 @@ func Test_SDSubmit_ValidateLogin_Pass_None(t *testing.T) {
 
 	if len(s.datasets) > 0 {
 		t.Errorf("Function failed, expected no datasets, received=%d", len(s.datasets))
-	}
-}
-
-func Test_SDSubmit_LevelCount(t *testing.T) {
-	s := sdSubmitInfo{}
-	lc := s.levelCount()
-	if lc != 2 {
-		t.Errorf("Function failed, expected=%d, received=%d", 2, lc)
 	}
 }